requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
python-dotenv==1.1.1
asyncpg